autofs-5.1.6 - improve sss getautomntbyname() error handling
Recent versions of the sss autofs access library will return EHOSTDOWN
if the backend server is down.

The presence of this improvement in error handling is determined by an
added function to get the sss autofs protocol version.

Update the getautomntbyname() function to use this.

Also introduce a wait function so we can wait (for a while) for the host
to come back up.

Signed-off-by: Ian Kent <raven@themaw.net>