autofs-5.1.6 - simplify sss source stale check
Now that there's a positive timeout on valid map cache entries
simplifying the map stale checks should be able to be done, and
eventually this will be much simpler in a lot of places.

Start with the key lookup in the sss lookup module since that's
needed to prevent recently retrieved map keys from being pruned
immediately after being used since the existing check wasn't
working and was questionable anyway (and likely hadn't for worked
some time).

Signed-off-by: Ian Kent <raven@themaw.net>