autofs-5.1.6 - make autofs.a a shared library
When the linker option -Bsymbolic-functions is used the static library
that is linked to each object, automount and the shared lookup, parse
and mount shared libraries becomes unusable because static storage
allocated and initialized in the the library becomes undefined after
certain events, such as forking to become a daemon.

This can be fixed by changing the autofs static library to a shared
library.