autofs-5.1.6 - fix remount expire
When starting autofs when there are active mounts due to in use
mounts from a previous shutdown, and a mount entry has offsets,
and an offset doesn't have a real mount due to a non-strict fail
returning success for the remount a file handle is incorrectly
left open.

The file handle gets assingned to the offset entry for expires
but since there is no mount to expire it never gets closed which
prevents the multi-mount owner itself from expiring.

Signed-off-by: Ian Kent <raven@themaw.net>