autofs-5.1.6 - fix stale offset directories disable mount
When starting autofs when there are active mounts due to in use
mounts from a previous shutdown, and a mount entry has offsets
that are within an autofs mount, the mount point directories are
incorrectly marked as not created by us when re-connecting to the
mount so they are not removed at expire.

Since the base autofs mount directory needs to be empty for mounts
to be triggered these directories disable automounting for the mount.

Signed-off-by: Ian Kent <raven@themaw.net>