autofs-5.1.7 - eliminate count_mounts() from expire_proc_indirect()
The count_mounts() function traverses the directory tree under a given
automount in order to count the number of mounts.

If there are many directories (such as when there is a very large
number of offset trigger mounts) this can take a long time.

Eliminate the call in expire_proc_indirect() by changing the expire
ioctl function to better use the expire return from the kernel.

Signed-off-by: Ian Kent <raven@themaw.net>