autofs-5.1.7 - reduce umount EBUSY check delay
Some time ago I had to wait and retry umount() for autofs mounts
becuase I found EBUSY would be returned for a time after the call
causing false negative umount returns.

I think that problem has been resolved but removing the retry is
probably a little risky.

But the wait time is quite long at one fifth of a second so reduce
that to one twentieth of a second and increase the retries to make
it more resposive.

Signed-off-by: Ian Kent <raven@themaw.net>