autofs-5.1.7 - make tree implementation data independent
Generalise the tree implementation so that it's independent of the
data structure that's used.

Do this by refactoring it into core tree functions and functions
specific to the data structure to be used so that different data
structures can be used when needed by adding an implementation for
the data structure specific functions.

Signed-off-by: Ian Kent <raven@themaw.net>