autofs-5.1.7 - fix a couple of null cache locking problems
There's no locking used for null cache access in mount_autofs_direct().
And in master_mount_mounts() an entry could be deleted holding the read
lock only.

Also in each of these cases an unnecessary cache_partial_match() is
done.

In do_readmap() the null cache read lock is taken but it is only needed
for a short time in do_readmap_mount() where an entry could be deleted.

Signed-off-by: Ian Kent <raven@themaw.net>