autofs-5.1.7 - fix amd hosts mount expire
When swicthing to use the mnt_list to track mounts for expire, if the
amd hosts map entry name is for the host short name, the amd mount
entry for the short name gets removed. This causes a subsequent mounts
for host exports to fail.

What should happen is the short name amd entry not be removed and a
mounted mount entry for the symlinked FQDN mount added so it expires.

Signed-off-by: Ian Kent <raven@themaw.net>