autofs-5.1.7 - eliminate redundant cache lookup in tree_mapent_add_node()
Since we need to create the offset tree after adding the offset entries
to the mapent cache (from a list.h list) there's no need to lookup the
mapent in tree_mapent_add_node() and validate it. Just use it directly
when calling tree_mapent_add_node() and avoid a cache lookup on every
node addition.

Signed-off-by: Ian Kent <raven@themaw.net>