autofs-5.1.7 - fix hosts map offset order
Map entry offset paths need to be in shortest to longest order but
exports from a server could come in any order. If there are a large
number of exports this can result in a lot of overhead when adding
the offset to the ordered list used to mount the offset during parsing
since the path length of exports can vary a lot.

So leverage the tree implemention to sort the export offsets into
shortest to longest order as we go when constructing the mapent from
the exports list.

Signed-off-by: Ian Kent <raven@themaw.net>