autofs-5.1.7 - fix nonstrict offset mount fail handling
If a triggered offset mount fails automount is not handling nonstrict
mount failure correctly.

The nonstrict mount failure handling needs to convert an offset mount
failure to a success if the offset subtree below the failed mount is not
empty otherwise it must return the failure. The previous implementation
used -1 to indicate the subtree was empty and that was used to detect
when the mount should fail instead of converting the fail to a success.

Make the new implementation do the same.

Signed-off-by: Ian Kent <raven@themaw.net>