autofs-5.1.7 - eliminate buffer usage from handle_mounts_cleanup()
This buffer was originally added because a SEGV was seen accessing
the ap->path field on shutdown.

But this was actually caused by calling master_remove_mapent() too
early which adds the map entry to the master map join list that leads
to freeing the autofs_point (ap in the code) which also frees ap->path.

But the master map join list is protected by the master map mutex which
is held until after all the accesses are completed. So whatever the
problem was it doesn't appear to be present any more.

Nevertheless, to be sure, delay the call to master_remove_mapent() until
after all accesses to ap->path are completed.

Signed-off-by: Ian Kent <raven@themaw.net>