autofs-5.1.7 - make NFS version check flags consistent
Several of the NFS connection macros have the same value so that they
can be used as internal code documentation of what is being done.

Adjust the protocol macro naming to be consistent in a few places.

Also make sure the correct flags are set for the function they indicate.

Signed-off-by: Ian Kent <raven@themaw.net>