autofs-5.1.8 - fix set open file limit
The check of whether the open file limit needs to be changed is not
right, it checks the hard open file limit against what autofs wants
to set it to which is always less than this value. Consequently the
open file limit isn't changed.

autofs should be changing only the soft open file limit but it is
setting both the hard and soft limits. The system hard limit is much
higer than the autofs maximum open files so the hard limit should be
left alone.

While we are here increase the requested maximum soft open file limit
to 20k.

Signed-off-by: Ian Kent <raven@themaw.net>