autofs-5.1.8 - dont use initgroups() at spawn
The initgroups(3) function isn't safe to use between fork() and
exec() in a threaded program.

Using it this way often leads to a hang for even moderate work
loads.

But the getgrouplist()/setgroups() combination can be used safely
in this case and this patch changes autofs to use these (the safety
of using of setgroups() is yet to to be documented).

A large portion of the work on this patch has been contributed
by Roberto Bergantinos <rbergant@redhat.com>.

Reported-by: Roberto Bergantinos <rbergant@redhat.com>
Fixes: 6343a3292020 ("autofs-5.1.3 - fix ordering of seteuid/setegid in do_spawn()")
Signed-off-by: Roberto Bergantinos <rbergant@redhat.com>
Signed-off-by: Ian Kent <raven@themaw.net>