autofs-5.1.8 - more comprehensive verbose logging for LDAP maps
Current logging of LDAP map lookups in verbose mode is not very
comprehensive. It's not clear for what purpose connections to the LDAP
directory are made as the only result that will be reported is the autofs
mount point creation.

Inform users about the intention of the LDAP directory access in verbose
mode.

Signed-off-by: Thomas Reim <reimth@gmail.com>
Signed-off-by: Ian Kent <raven@themaw.net>