autofs-5.1.8 - fix hosts map deadlock on restart
When starting automount(8) with a hosts map that has mounts that were
in use at the last exit a deadlock can occur.

In this case automount(8) will perform the same steps but not actually
perform the mount to re-construct the context of each mount. But, with
the hosts map, that leads to calling back into the sun parse module
while holding the map module read lock which will again try and take
the write lock.

Fix this by only taking the write lock in the mount code path if the
module handle has not already been opened.

Signed-off-by: Ian Kent <raven@themaw.net>