autofs-5.1.8 - fix deadlock with hosts map reload
When reloading maps the hosts map calls lookup method ->parse_mount()
for each multi-mount root entry in the map (each host) while holding
the cache read lock which leads to a cache lock deadlock.

Remove the need to hold the cache read lock by creating an independent
list of entries for the update so the lock doesn't need to be taken.

Signed-off-by: Ian Kent <raven@themaw.net>