autofs-5.1.8 - use device id to locate autofs_point when setting log priotity
Using a fifo pipe for every autofs mount to dynamically set the log
priority is expensive in terms of the number of file handles used.

It would be better to use a single file handle and locate the autofs
mount point by it's id to set the log priority.

Start by making the communication pipe send the device id as well as
the log priority to be set and use the newly added helper function
master_find_mapent_by_devid() to locate the autofs mount to change the
log priority.

Signed-off-by: Ian Kent <raven@themaw.net>