[PATCH] autofs-5.1.8 - define LDAP_DEPRECATED during LDAP configure check
This commit defines LDAP_DEPRECATED as 1 during a configure check for
ldap_parse_page_control.  This is in line with how lookup_ldap.c is
compiled at build time.

The configure script is regenerated with autoconf 2.69.

The regeneration of the configure script also includes changes due to
commit 69fda4f090e3.  That commit intended to run checks with implicit
function declaration warnings enabled in order to recognize missing
libldap functions.  However, the in-tree copy of the configure script
was not regenerated at that time.

Signed-off-by: Arjun Shankar <arjun@redhat.com>
Signed-off-by: Ian Kent <raven@themaw.net>