autofs-5.1.8 - don't close lookup at umount
Since map sources are reference counted they persist beyond autofs
submounts.

Now the map source moudule lookup gets closed at submount umount and
if we are unlucky enough to be using the same map in other submounts
and a lookup is underway at the time of the umount a crash can occur.

To resolve this it's much better to just not close the lookup at
submount umount and rely on the map source free to close the module
lookup and instances when the map source is no longer referenced.

Signed-off-by: Ian Kent <raven@themaw.net>