autofs-5.1.8 - fix incorrect matching of cached wildcard key
During the implementation of amd format map entry support the code
to match a cached key was modified.

Unfortunately there's a case were the key lookup behaves incorrectly.

That case is when there are included maps in the map itself and one
of the maps (usually the last) has a wildcard key entry. In this case
the wildcard key may be found during lookup but the map it blongs to
isn't checked so it can be incorrectly returned instead of a matching
entry in a subsequent included map.

Another problem case is when there's a wildcard match and a cache prune
occurs while the mount is being done. In this case the matched cache
entry that has been added is seen as stale and removed along with the
mount point directory during the prune leading to a mount fail.

Signed-off-by: Ian Kent <raven@themaw.net>