autofs-5.1.8 - dont probe interface that cant send packet
When calculating the proximity add checks for basic reachability.

If an interface doesn't have an address of the family of the target
host, or the interface address is the IPv6 link local address, or
the target host address is the IPv6 link local address then don't
add it to the list of hosts to probe.

Reported-by: Salvatore Bonaccorso <carnil@debian.org>
Tested-by: Salvatore Bonaccorso <carnil@debian.org>
Cc: Goldwyn Rodrigues <rgoldwyn@suse.de>
Cc: Mike Gabriel <sunweaver@debian.org>
Signed-off-by: Ian Kent <raven@themaw.net>