autofs-5.1.9 - refactor amd function umount_amd_ext_mount()
The amd mounts function umount_amd_ext_mount() needs some improvement.

Make sure the function returns true for success and false for failure
and add a parameter to control if the expternal mount reference should
be decremented on successful umount.

If the reference count of the external mount is greater than 1 there's
some other mount using (symlink pointing to) it so don't try to umount
it just return success.

Also check for the case where the mount is already mounted.

Signed-off-by: Ian Kent <raven@themaw.net>