autofs-5.1.9 - add flags argument to amd do_program_mount()
Most of the amd mount functions take a flags argument that allows them
to alter their function based on configuration.

For example the amd option autofs_use_lofs will use bind mounts instead
of symlinks in some cases which might be preferred.

The program mount function was not being passed this parameter but the
design of all the amd mount functions is quite similar and adding the
flag works as expected..

Signed-off-by: Ian Kent <raven@themaw.net>